14 Reflection
Describe the differences between forks, clones, and branches. When would you use one instead of another?
The differences between forks, clones, and branches are that forks give credit to the original creator, clones create an entirely new group of code while the other one is still present, and branches can be merged back with the master branch.